rename cupy->loadcupy and fix jastrow updateinternals masked assignment error
